refactoring to merge
