+++ com/walkroute/llist/RouteNode.java
@@ -1,4 +1,4 @@
-package walkroute;
+package com.walkroute.llist;
 
 public class RouteNode {
     String location;
